docs/5-ml-fundamentals/5.4-unsupervised-learning/requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
umap-learn>=0.5.0
scipy>=1.7.0 
